# internal/adapters/http/docs/swagger.yaml
basePath: /api
definitions:
...
  response.DataResponse-role_CreateResultDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/role.CreateResultDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-role_RoleDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/role.RoleDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.MessageResponse:
    description: 纯消息响应格式
    properties:
      code:
        description: HTTP 状态码
        type: integer
      message:
        description: 消息描述
        type: string
    type: object
...
  response.PagedResponse-role_RoleDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/role.RoleDTO'
        type: array
      message:
        description: 消息描述
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: 分页信息
    type: object
...
  response.PaginationMeta:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      per_page:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
      warning:
        description: 页码越界警告
        type: string
    type: object
  role.CreateDTO:
    properties:
      description:
        example: 系统开发人员角色
        maxLength: 255
        type: string
      display_name:
        example: 开发者
        maxLength: 100
        type: string
      name:
        example: developer
        maxLength: 50
        minLength: 2
        type: string
    required:
    - display_name
    - name
    type: object
  role.CreateResultDTO:
    properties:
      display_name:
        type: string
      name:
        type: string
      role_id:
        type: integer
    type: object
  role.PermissionDTO:
    properties:
      operation_pattern:
        example: sys:users.*
        type: string
      resource_pattern:
        example: user/*
        type: string
    type: object
  role.PermissionInputDTO:
    properties:
      operation_pattern:
        example: sys:users.*
        type: string
      resource_pattern:
        example: user/*
        type: string
    required:
    - operation_pattern
    type: object
  role.RoleDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/role.PermissionDTO'
        type: array
      updated_at:
        type: string
    type: object
  role.SetPermissionsDTO:
    properties:
      permissions:
        items:
          $ref: '#/definitions/role.PermissionInputDTO'
        type: array
    required:
    - permissions
    type: object
  role.UpdateDTO:
    properties:
      description:
        maxLength: 255
        type: string
      display_name:
        maxLength: 100
        type: string
    type: object
...
host: localhost:40012
info:
  contact:
    email: lwmacct@icloud.com
    name: API Support
    url: https://github.com/lwmacct/251117-go-ddd-template/issues
  description: 基于 DDD+CQRS 架构的 Go Web 应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lwmacct/251117-go-ddd-template
  title: Go DDD Template API
  version: 1.0.3
paths:
...
  /api/admin/roles:
    get:
      consumes:
      - application/json
      description: 分页获取所有系统角色
      parameters:
      - default: 20
        description: Limit 每页数量，默认 20，最大 1000
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page 页码，从 1 开始
        in: query
        minimum: 1
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色列表
          schema:
            $ref: '#/definitions/response.PagedResponse-role_RoleDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 角色列表
      tags:
      - Admin - Role Management
    post:
      consumes:
      - application/json
      description: 管理员创建新的系统角色
      parameters:
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.CreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 角色创建成功
          schema:
            $ref: '#/definitions/response.DataResponse-role_CreateResultDTO'
        "400":
          description: 参数错误或角色名已存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - Admin - Role Management
  /api/admin/roles/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定角色（如果角色被用户使用，可能会失败）
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色删除成功
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: 无效的角色ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误或角色被使用中
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - Admin - Role Management
    get:
      consumes:
      - application/json
      description: 根据角色ID获取角色详细信息（包含权限列表）
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色详情
          schema:
            $ref: '#/definitions/response.DataResponse-role_RoleDTO'
        "400":
          description: 无效的角色ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 角色详情
      tags:
      - Admin - Role Management
    put:
      consumes:
      - application/json
      description: 管理员更新角色的显示名称和描述
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.UpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 角色更新成功
          schema:
            $ref: '#/definitions/response.DataResponse-role_RoleDTO'
        "400":
          description: 无效的角色ID或参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - Admin - Role Management
  /api/admin/roles/{id}/permissions:
    put:
      consumes:
      - application/json
      description: 管理员为指定角色设置权限（会覆盖现有权限）。新 RBAC 模型使用 Operation + Resource Pattern。
      parameters:
      - description: 角色ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 权限模式列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.SetPermissionsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 权限设置成功
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: 无效的角色ID或参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设置权限
      tags:
      - Admin - Role Management
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"